requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
uuid==1.30
wrapt==1.11.2
